ref: update fee
