+++ contracts/hub/stable_vlp/src/contract.rs
@@ -2,23 +2,21 @@ use std::collections::HashMap;
 
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
-use cosmwasm_std::{Binary, Deps, DepsMut, Env, MessageInfo, Reply, Response, Uint128, Uint64};
+use cosmwasm_std::{Binary, Deps, DepsMut, Env, MessageInfo, Reply, Response, Uint128};
 use cw2::set_contract_version;
 use euclid::fee::{DenomFees, TotalFees};
 
-use crate::execute::{
-    add_liquidity, execute_swap, register_pool, remove_liquidity, update_fee, update_state,
-};
-use crate::reply::{NEXT_SWAP_REPLY_ID, VIRTUAL_BALANCE_TRANSFER_REPLY_ID};
-use crate::state::{State, AMP_FACTOR, BALANCES, DEFAULT_AMP_FACTOR, STATE};
-use crate::{execute, reply};
-use euclid::error::ContractError;
-use euclid::msgs::stable_vlp::{ExecuteMsg, InstantiateMsg, QueryMsg};
-
+use crate::execute::{add_liquidity, execute_swap, register_pool, remove_liquidity, update_state};
 use crate::query::{
     query_all_pools, query_fee, query_liquidity, query_pool, query_simulate_swap, query_state,
     query_total_fees_collected, query_total_fees_per_denom,
 };
+use crate::reply::{NEXT_SWAP_REPLY_ID, VIRTUAL_BALANCE_TRANSFER_REPLY_ID};
+use crate::state::{AMP_FACTOR, BALANCES, DEFAULT_AMP_FACTOR, STATE};
+use crate::{execute, reply};
+use euclid::error::ContractError;
+use euclid::msgs::stable_vlp::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use euclid::pool::{update_fee, State};
 // version info for migration info
 const CONTRACT_NAME: &str = "crates.io:stable_vlp";
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -94,7 +92,7 @@ pub fn execute(
             lp_fee_bps,
             euclid_fee_bps,
             recipient,
-        } => update_fee(deps, info, lp_fee_bps, euclid_fee_bps, recipient),
+        } => update_fee(deps, info, &STATE, lp_fee_bps, euclid_fee_bps, recipient),
         ExecuteMsg::AddLiquidity {
             sender,
             tx_id,
