+++ contracts/hub/stable_vlp/src/execute.rs
@@ -1,12 +1,19 @@
+use crate::{
+    math::compute_swap,
+    query::{assert_slippage_tolerance, calculate_lp_allocation},
+    reply::{NEXT_SWAP_REPLY_ID, VIRTUAL_BALANCE_TRANSFER_REPLY_ID},
+    state::{self, AMP_FACTOR, BALANCES, CHAIN_LP_TOKENS, DEFAULT_AMP_FACTOR, STATE},
+};
 use cosmwasm_std::{
     ensure, to_json_binary, Decimal, Decimal256, DepsMut, Env, MessageInfo, Response, SubMsg,
     Uint128, Uint64, WasmMsg,
 };
+use euclid::pool::State;
 use euclid::{
     chain::{ChainUid, CrossChainUser},
     error::ContractError,
     events::{liquidity_event, simple_event, tx_event, TxType},
-    fee::{Fee, BPS_50_PERCENT, MAX_FEE_BPS},
+    fee::{Fee, BPS_50_PERCENT},
     liquidity::AddLiquidityResponse,
     msgs::{
         stable_vlp::{VlpRemoveLiquidityResponse, VlpSwapResponse},
@@ -16,13 +23,6 @@ use euclid::{
     swap::NextSwapVlp,
     token::{Pair, PairWithAmount, Token},
     utils::math::Decimal256Ext,
-};
-
-use crate::{
-    math::compute_swap,
-    query::{assert_slippage_tolerance, calculate_lp_allocation},
-    reply::{NEXT_SWAP_REPLY_ID, VIRTUAL_BALANCE_TRANSFER_REPLY_ID},
-    state::{self, State, AMP_FACTOR, BALANCES, CHAIN_LP_TOKENS, DEFAULT_AMP_FACTOR, STATE},
 };
 
 /// Registers a new pool in the contract. Function called by Router Contract
@@ -628,35 +628,6 @@ pub fn execute_swap(
         .add_attribute("lp_fee", lp_fee)
         .add_attribute("receive_amount", receive_amount)
         .set_data(acknowledgement))
-}
-
-pub fn update_fee(
-    deps: DepsMut,
-    info: MessageInfo,
-    lp_fee_bps: Option<u64>,
-    euclid_fee_bps: Option<u64>,
-    recipient: Option<CrossChainUser>,
-) -> Result<Response, ContractError> {
-    let mut state = STATE.load(deps.storage)?;
-    ensure!(info.sender == state.admin, ContractError::Unauthorized {});
-
-    state.fee.lp_fee_bps = lp_fee_bps.unwrap_or(state.fee.lp_fee_bps);
-    ensure!(
-        state.fee.lp_fee_bps.le(&MAX_FEE_BPS),
-        ContractError::new("LP Fee cannot exceed maximum limit")
-    );
-    state.fee.euclid_fee_bps = euclid_fee_bps.unwrap_or(state.fee.euclid_fee_bps);
-    ensure!(
-        state.fee.euclid_fee_bps.le(&MAX_FEE_BPS),
-        ContractError::new("Euclid Fee cannot exceed maximum limit")
-    );
-    state.fee.recipient = recipient.unwrap_or(state.fee.recipient);
-
-    STATE.save(deps.storage, &state)?;
-
-    Ok(Response::new()
-        .add_event(simple_event())
-        .add_attribute("action", "update_fee"))
 }
 
 pub fn update_state(
