+++ contracts/hub/stable_vlp/src/tests.rs
@@ -3,13 +3,14 @@
 mod tests {
     use crate::contract::{execute, instantiate};
     use crate::math::compute_swap;
-    use crate::state::{State, BALANCES, CHAIN_LP_TOKENS, STATE};
+    use crate::state::{BALANCES, CHAIN_LP_TOKENS, STATE};
     use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
     use cosmwasm_std::{coins, Decimal256, DepsMut, Response, Uint128, Uint64};
     use euclid::chain::{ChainUid, CrossChainUser};
     use euclid::error::ContractError;
     use euclid::fee::{DenomFees, Fee, TotalFees};
     use euclid::msgs::stable_vlp::{ExecuteMsg, InstantiateMsg};
+    use euclid::pool::State;
     use euclid::token::{Pair, Token};
     use std::collections::HashMap;
