+++ contracts/hub/vlp/src/contract.rs
@@ -6,14 +6,13 @@ use cosmwasm_std::{Binary, Deps, DepsMut, Env, MessageInfo, Reply, Response, Uin
 use cw2::set_contract_version;
 use euclid::fee::{DenomFees, TotalFees};
 
-use crate::execute::{
-    add_liquidity, execute_swap, register_pool, remove_liquidity, update_fee, update_state,
-};
+use crate::execute::{add_liquidity, execute_swap, register_pool, remove_liquidity, update_state};
 use crate::reply::NEXT_SWAP_REPLY_ID;
-use crate::state::{State, BALANCES, STATE};
+use crate::state::{BALANCES, STATE};
 use crate::{execute, reply};
 use euclid::error::ContractError;
 use euclid::msgs::vlp::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use euclid::pool::{update_fee, State};
 
 use crate::query::{
     query_all_pools, query_fee, query_liquidity, query_pool, query_simulate_swap, query_state,
@@ -92,7 +91,7 @@ pub fn execute(
             lp_fee_bps,
             euclid_fee_bps,
             recipient,
-        } => update_fee(deps, info, lp_fee_bps, euclid_fee_bps, recipient),
+        } => update_fee(deps, info, &STATE, lp_fee_bps, euclid_fee_bps, recipient),
         ExecuteMsg::AddLiquidity {
             sender,
             tx_id,
