+++ packages/euclid/src/pool.rs
@@ -1,9 +1,13 @@
 use crate::{
     chain::CrossChainUser,
-    token::{PairWithDenomAndAmount, TokenWithDenom},
+    error::ContractError,
+    events::simple_event,
+    fee::{Fee, TotalFees, MAX_FEE_BPS},
+    token::{Pair, PairWithDenomAndAmount, TokenWithDenom},
 };
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Uint128, Uint64};
+use cosmwasm_std::{ensure, DepsMut, MessageInfo, Response, Uint128, Uint64};
+use cw_storage_plus::Item;
 
 pub const MINIMUM_LIQUIDITY: u128 = 1000;
 
@@ -62,19 +66,53 @@ pub enum PoolConfig {
     ConstantProduct {},
 }
 
-// #[cw_serde]
-// pub struct PoolInstantiateMsg {
-//     pub router: String,
-//     pub virtual_balance: String,
-//     pub pair: Pair,
-//     pub fee: Fee,
-//     pub execute: Option<PoolExecuteMsg>,
-//     pub admin: String,
-//     pub amp_factor: Option<Uint64>,
-// }
+#[cw_serde]
+pub struct State {
+    // Token Pair Info
+    pub pair: Pair,
+    // Router Contract
+    pub router: String,
+    // Virtual Coin Contract
+    pub virtual_balance: String,
+    // Fee per swap for each transaction
+    pub fee: Fee,
+    // Total lp and euclid fees collected
+    pub total_fees_collected: TotalFees,
+    // The last timestamp where the balances for each token have been updated
+    pub last_updated: u64,
+    // total number of LP tokens issued
+    pub total_lp_tokens: Uint128,
+    pub admin: String,
+}
 
-// #[cw_serde]
-// pub enum PoolExecuteMsg {
-//     Stable(StableExecuteMsg),
-//     ConstantProduct(ConstantProductExecuteMsg),
-// }
+pub fn update_fee(
+    deps: DepsMut,
+    info: MessageInfo,
+    state_storage: &Item<State>, // Pass as a reference
+    lp_fee_bps: Option<u64>,
+    euclid_fee_bps: Option<u64>,
+    recipient: Option<CrossChainUser>,
+) -> Result<Response, ContractError> {
+    let mut state = state_storage.load(deps.storage)?;
+    ensure!(info.sender == state.admin, ContractError::Unauthorized {});
+
+    state.fee.lp_fee_bps = lp_fee_bps.unwrap_or(state.fee.lp_fee_bps);
+    ensure!(
+        state.fee.lp_fee_bps <= MAX_FEE_BPS,
+        ContractError::new("LP Fee cannot exceed maximum limit")
+    );
+
+    state.fee.euclid_fee_bps = euclid_fee_bps.unwrap_or(state.fee.euclid_fee_bps);
+    ensure!(
+        state.fee.euclid_fee_bps <= MAX_FEE_BPS,
+        ContractError::new("Euclid Fee cannot exceed maximum limit")
+    );
+
+    state.fee.recipient = recipient.unwrap_or(state.fee.recipient);
+
+    state_storage.save(deps.storage, &state)?;
+
+    Ok(Response::new()
+        .add_attribute("action", "update_fee")
+        .add_event(simple_event()))
+}
